+++ src/main/java/xbot/common/controls/actuators/wpi_adapters/CANdleWpiAdapter.java
@@ -144,7 +144,7 @@ public class CANdleWpiAdapter extends XCANLightController {
     }
 
     @Override
-    public void twinkleOff(int slot, Frequency frameRate, int density, Color color) {
+    public void twinkleOff(int slot, Frequency frameRate, double density, Color color) {
         this.candle.setControl(
                 new TwinkleOffAnimation(getSlotStartIndex(slot), getSlotEndIndex(slot))
                         .withSlot(slot)
